+++ function.h
@@ -2,4 +2,4 @@
 int calc(int, int);//Adding
 int calc(short, short);//Substract
 int calc(double, double);//Multiply
-int calc(float , float);//Divide
+float calc(float , float);//Divide
